requirements.txt
PyGithub==1.59.1
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
pyyaml==6.0.1
requests==2.31.0
jsonschema==4.19.0
python-dotenv==1.0.0
